create course and course show

+++ course/del_course.php
@@ -5,9 +5,18 @@ include_once '../config.php';
 if(isset($_GET['id'])){
     $id = $_GET['id'];
 
-    $sql = "DELETE FROM course WHERE id='$id'";
-    $result = mysqli_query($conn,$sql);
+    $sql1 = "SELECT * FROM course WHERE id = '$id'";
+    $result1 = mysqli_query($conn,$sql1);
+    $row = mysqli_fetch_assoc($result1);
 
+    $location = 'upload/'.$row['image'].''; //ตัวแปรที่อยู่ของภาพ
+
+    if(unlink($location)){ //unlink คือ การลบรูปออกจากโฟเดอร์ที่จัดเก็บอยู่
+
+        $sql = "DELETE FROM course WHERE id='$id'";
+        $result = mysqli_query($conn,$sql);
+    }
+    
     if($result){
         $_SESSION['success'] = "ลบข้อมูลสำเร็จ";
         header( "location: ../admin.php?Menu=2&Submenu=indexco" );
